split query
